requirements.txt
aerich
aiogram==3.20.0
aiohttp==3.11.0
aiosqlite
cachetools
coloredlogs
tgcrypto
toml
tortoise-orm
PyGithub
fastapi
uvicorn
coloredlogs